Excerise week 1 day 3

+++ module-1/03_Expressions/student-breakouts/01_ExpressionsBreakOut/Lecture/04_ReturnDoubleOfTwo.cs
@@ -12,7 +12,7 @@ namespace Lecture
          4. This method currently returns an int. Change it so that it returns a double.
             TOPIC: Return Types
          */
-        public int ReturnDoubleOfTwo()
+        public double ReturnDoubleOfTwo()
         {
             return 2;
         }
